+++ backend/posts/apps.py
@@ -4,3 +4,6 @@ from django.apps import AppConfig
 class PostsConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
     name = 'posts'
+    
+    def ready(self) -> None:
+        import posts.signals
